+++ server/src/main/java/org/opensearch/rest/RestController.java
@@ -35,7 +35,12 @@ package org.opensearch.rest;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
+import org.apache.shiro.authc.AuthenticationException;
 import org.opensearch.OpenSearchException;
+import org.opensearch.authn.tokens.AuthenticationToken;
+import org.opensearch.authn.tokens.BasicAuthToken;
+import org.opensearch.authn.tokens.HttpHeaderToken;
+import org.opensearch.authn.Subject;
 import org.opensearch.client.node.NodeClient;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.Strings;
@@ -50,6 +55,7 @@ import org.opensearch.common.xcontent.XContentBuilder;
 import org.opensearch.common.xcontent.XContentType;
 import org.opensearch.core.internal.io.Streams;
 import org.opensearch.http.HttpServerTransport;
+import org.opensearch.identity.Identity;
 import org.opensearch.indices.breaker.CircuitBreakerService;
 import org.opensearch.usage.UsageService;
 
@@ -57,11 +63,13 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Supplier;
@@ -395,6 +403,9 @@ public class RestController implements HttpServerTransport.Dispatcher {
                         return;
                     }
                 } else {
+                    // Authenticate incoming request
+                    if (!authenticate(request, channel)) return;
+
                     dispatchRequest(request, channel, handler);
                     return;
                 }
@@ -587,5 +598,73 @@ public class RestController implements HttpServerTransport.Dispatcher {
     private static CircuitBreaker inFlightRequestsBreaker(CircuitBreakerService circuitBreakerService) {
         // We always obtain a fresh breaker to reflect changes to the breaker configuration.
         return circuitBreakerService.getBreaker(CircuitBreaker.IN_FLIGHT_REQUESTS);
+    }
+
+    /**
+     * Authenticates the subject of the incoming REST request based on the auth header
+     * @param request the request whose subject is to be authenticated
+     * @param channel the channel to send the response on
+     * @return true if authentication was successful, false otherwise
+     * @throws IOException when an exception is raised writing response to channel
+     */
+    private boolean authenticate(RestRequest request, RestChannel channel) throws IOException {
+
+        final Optional<String> authHeader = request.getHeaders()
+            .getOrDefault(HttpHeaderToken.HEADER_NAME, Collections.emptyList())
+            .stream()
+            .findFirst();
+
+        Subject subject = null;
+
+        AuthenticationToken headerToken = null;
+
+        if (authHeader.isPresent()) {
+            try {
+                headerToken = tokenType(authHeader.get());
+                subject = Identity.getAuthManager().getSubject();
+                subject.login(headerToken);
+                logger.info("Authentication successful");
+                return true;
+            } catch (final AuthenticationException ae) {
+                logger.info("Authentication finally failed: {}", ae.getMessage());
+
+                final BytesRestResponse bytesRestResponse = BytesRestResponse.createSimpleErrorResponse(
+                    channel,
+                    RestStatus.UNAUTHORIZED,
+                    ae.getMessage()
+                );
+                channel.sendResponse(bytesRestResponse);
+                return false;
+            }
+        }
+
+        // TODO: Handle anonymous Auth - Allowed or Disallowed (set by the user of the system) - 401 or Login-redirect ??
+
+        /*
+        TODO: Uncomment this once it is decided to proceed with this workflow
+        logger.info("Authentication unsuccessful: Missing Authentication Header");
+        final BytesRestResponse bytesRestResponse = BytesRestResponse.createSimpleErrorResponse(
+            channel,
+            RestStatus.BAD_REQUEST,
+            "Missing Authentication Header"
+        );
+        channel.sendResponse(bytesRestResponse);
+        */
+
+        // This is allowing headers without Auth header to pass through.
+        // At the time of writing this, all rest-tests would fail if this is set to false
+        // TODO: Change this to false once there is a decision on what to do with requests that don't have auth Headers
+        return true;
+    }
+
+    /**
+     * Identifies the token type and return the correct instance
+     * @param authHeader from which to identify the correct token class
+     * @return the instance of the token type
+     */
+    private AuthenticationToken tokenType(String authHeader) {
+        if (authHeader.contains("Basic")) return new BasicAuthToken(authHeader);
+        // support other type of header tokens
+        return null;
     }
 }