+++ server/src/test/java/org/opensearch/rest/RestControllerTests.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.rest;
 
+import com.carrotsearch.randomizedtesting.annotations.ThreadLeakScope;
 import org.opensearch.client.node.NodeClient;
 import org.opensearch.common.breaker.CircuitBreaker;
 import org.opensearch.common.bytes.BytesArray;
@@ -52,6 +53,9 @@ import org.opensearch.http.HttpRequest;
 import org.opensearch.http.HttpResponse;
 import org.opensearch.http.HttpServerTransport;
 import org.opensearch.http.HttpStats;
+import org.opensearch.authn.AuthenticationManager;
+import org.opensearch.identity.Identity;
+import org.opensearch.authn.internal.InternalAuthenticationManager;
 import org.opensearch.indices.breaker.HierarchyCircuitBreakerService;
 import org.opensearch.test.OpenSearchTestCase;
 import org.opensearch.test.client.NoOpNodeClient;
@@ -84,6 +88,7 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+@ThreadLeakScope(ThreadLeakScope.Scope.NONE) /* otherwise {@code testRestRequestAuthentication} cause thread to be leaked */
 public class RestControllerTests extends OpenSearchTestCase {
 
     private static final ByteSizeValue BREAKER_LIMIT = new ByteSizeValue(20);
@@ -636,6 +641,42 @@ public class RestControllerTests extends OpenSearchTestCase {
             channel.getRestResponse().content().utf8ToString(),
             equalTo("{\"error\":\"Unexpected HTTP method, allowed: [GET]\",\"status\":405}")
         );
+    }
+
+    // Tests to check authenticate(...) method
+    public void testRestRequestAuthenticationSuccess() {
+        final AuthenticationManager authManager = new InternalAuthenticationManager();
+        Identity.setAuthManager(authManager);
+
+        final ThreadContext threadContext = client.threadPool().getThreadContext();
+
+        final FakeRestRequest fakeRestRequest = new FakeRestRequest.Builder(NamedXContentRegistry.EMPTY).withHeaders(
+            Collections.singletonMap("Authorization", Collections.singletonList("Basic YWRtaW46YWRtaW4="))
+        ) // admin:admin
+            .build();
+        final AssertingChannel channel = new AssertingChannel(fakeRestRequest, true, RestStatus.OK);
+        restController.dispatchRequest(fakeRestRequest, channel, threadContext);
+
+        assertTrue(channel.getSendResponseCalled());
+    }
+
+    public void testRestRequestAuthenticationFailure() {
+        final AuthenticationManager authManager = new InternalAuthenticationManager();
+        Identity.setAuthManager(authManager);
+
+        final ThreadContext threadContext = client.threadPool().getThreadContext();
+
+        final FakeRestRequest fakeRestRequest = new FakeRestRequest.Builder(NamedXContentRegistry.EMPTY).withHeaders(
+            Collections.singletonMap("Authorization", Collections.singletonList("Basic bWFydmluOmdhbGF4eQ=="))
+        ) // marvin:galaxy
+            .build();
+
+        // RestStatus is OK even though the authn information is incorrect. This is because we, yet, don't fail the request
+        // if it was unauthorized. The status should be changed to UNAUTHORIZED once the flow is updated.
+        final AssertingChannel channel = new AssertingChannel(fakeRestRequest, true, RestStatus.UNAUTHORIZED);
+        restController.dispatchRequest(fakeRestRequest, channel, threadContext);
+
+        assertTrue(channel.getSendResponseCalled());
     }
 
     private static final class TestHttpServerTransport extends AbstractLifecycleComponent implements HttpServerTransport {